atualizando e adicionando metodos

+++ demo/src/main/java/com/omegascar/demo/repository/CarroRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarroRepository extends JpaRepository <Carro, Integer> {
+public interface CarroRepository extends JpaRepository<Carro, Long> {
     Carro findById(long id);
+
 }
